requirements.txt
slack-bolt
slack-sdk
openai
tiktoken
# https://github.com/Yelp/elastalert/issues/2306
urllib3<2
python-dotenv
flask
pyyaml
chromadb
langchain
requests
tabulate
unstructured
PyMuPDF